+++ src/components/Thumbnail/Thumbnail.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Image from 'next/image';
-import NoImage from '@/public/no-avatar.svg';
+import NoImage from '@/assets/images/no-avatar.svg';
 
 type Props = {
   imgUrl?: string;
