+++ src/pages/_app.tsx
@@ -1,4 +1,4 @@
-import '@/styles/globals.css';
+import '@/assets/styles/globals.css';
 import type { AppProps } from 'next/app';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { Raleway } from '@next/font/google';
